Refactored bookmark to allow better matching
Bookmarks now take regex parameters for regions that it matches. It
checks the text before the match and the text after the match before it
decides to link the match back to the bookmark.

The reasoning for this is that for small words that may have a lot of
false positives, we can tighten it up what it matches based on the regex
rules specified.

For example, the word ion will match function/condition/probation/etc..
But, if we set the bookmark to ensure that the text before ion doesn't
end with a letter/number AND the text after ion doesn't start with a
letter/number, it'll greatly reduce the number of false positives.